Moved Collection to top level namespace Updated settings to use Collection instead of a raw array
