+++ Slim/Collection.php
@@ -6,10 +6,10 @@
  * @copyright Copyright (c) 2011-2015 Josh Lockhart
  * @license   https://github.com/slimphp/Slim/blob/3.x/LICENSE.md (MIT License)
  */
-namespace Slim\Http;
+namespace Slim;
 
 use ArrayIterator;
-use Slim\Interfaces\Http\CollectionInterface;
+use Slim\Interfaces\CollectionInterface;
 
 /**
  * Collection
